Film/requirements.txt
Django==2.0.10
django-cors-headers==2.4.0
djangorestframework==3.9.0
pytz==2018.9
